case #10326

+++ Cloudents.Web/ClientApp/store/question.js
@@ -1,17 +1,28 @@
 import questionService from '../services/questionService'
 const state={
- correctAnswer:""
+ correctAnswer:"",
+    deletedAnswer:false
 }
 const mutations={
     updateAnswer(state,data){
         state.correctAnswer=data;
+    },
+    updateDeleted(state,data){
+        state.deletedAnswer=data;
     }
 }
 const getters={
-    getCorrectAnswer:(state)=>state.correctAnswer
+    getCorrectAnswer:(state)=>state.correctAnswer,
+    isDeletedAnswer:(state)=>state.deletedAnswer
 }
 const actions = {
+    resetQuestion({commit}){
+        commit('updateDeleted',false);
+        commit('updateAnswer','');
+
+    },
     deleteQuestion(context,id){
+       if(id.type==='Answer') {context.commit('updateDeleted',true)}
         return questionService.deleteQuestion(id);
     },
     correctAnswer(context,id){
